+++ src/main/java/com/aptzip/common/config/redis/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.aptzip.common.config;
+package com.aptzip.common.config.redis;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
